Ready for PR - just a suggestion
